requirements.txt
# Core LangChain + LangGraph + Google GenAI integration
langchain==1.0.3
langchain-core==1.0.2
langchain-community==0.2.9
langchain-chroma==0.1.4
langchain-google-genai==3.0.0
langgraph==1.0.2
langsmith==0.4.38

# Environment & HTTP
python-dotenv==1.0.0
requests==2.31.0

# Data models & tokenizers
pydantic>=2.0
tiktoken==0.4.0

# Transformers & embeddings
transformers==4.34.0
sentence-transformers==2.2.2
torch>=2.0.0
accelerate>=0.21.0

# NLP helpers & templating
nltk==3.8.1
jinja2==3.1.2

# HTML parsing and content handling
beautifulsoup4==4.12.2

# Vector DB
chromadb==1.3.3

# Utilities (logging, typing, etc.)
typing_extensions==4.8.0
email-validator==2.1.0.post1

# Optional: Data handling & file utilities
pandas==2.1.2
openpyxl==3.1.2

# Dev tools & testing (optional but recommended)
pytest==7.4.3
black==23.10.1
 
fastapi==0.121.1
prometheus-fastapi-instrumentator==7.1.0
